Rewards/ViewReward: add status information
Fixes #1629
