+++ apps/rewards/app/components/App/App.js
@@ -183,10 +183,14 @@ class App extends React.Component {
     })
   }
 
-  viewReward = (reward, panelTitle) => {
+  viewReward = (reward, isMyReward) => {
+    const panelTitle = isMyReward ?  'View my reward' : 'View Reward'
     this.setState({
       panel: PANELS.ViewReward,
-      panelProps: reward,
+      panelProps: {
+        reward,
+        isMyReward,
+      },
       panelTitle,
     })
   }
@@ -265,14 +269,14 @@ class App extends React.Component {
           <MyRewards
             myRewards={this.props.myRewards}
             myMetrics={this.props.myMetrics}
-            viewReward={reward => this.viewReward(reward, 'View my reward')}
+            viewReward={reward => this.viewReward(reward, true)}
             claimReward={this.claimReward}
           />
         ) : (
           <Overview
             rewards={this.props.rewards === undefined ? [] : this.props.rewards}
             newReward={this.newReward}
-            viewReward={reward => this.viewReward(reward, 'View Reward')}
+            viewReward={reward => this.viewReward(reward, false)}
             metrics={this.props.metrics}
           />
         )}
