+++ apps/rewards/app/components/Content/MyRewards.js
@@ -5,10 +5,6 @@ import {
   ContextMenu,
   ContextMenuItem,
   DataView,
-  GU,
-  IconCheck,
-  IconCircleCheck,
-  IconClock,
   IconCoin,
   IconView,
   Text,
@@ -23,37 +19,7 @@ import { Empty } from '../Card'
 import Metrics from './Metrics'
 import { useAppState } from '@aragon/api-react'
 import BigNumber from 'bignumber.js'
-import { displayCurrency } from '../../utils/helpers'
-
-const types = {
-  claimed: { icon: IconCircleCheck, text: 'Claimed', color: 'positive' },
-  ready: { icon: IconCheck, text: 'Ready to claim', color: 'positive' },
-  pending: { icon: IconClock, text: 'Pending', color: 'warning' },
-}
-
-const Status = ({ type }) => {
-  const theme = useTheme()
-  const status = types[type]
-  const text = status.text
-  const Icon = status.icon
-  const color = theme[status.color]
-  return (
-    <div css='display: flex;'>
-      <Icon style={{
-        marginRight: 0.5 * GU + 'px',
-        marginTop: -0.25 * GU + 'px',
-        color: color,
-      }} />
-      <Text>
-        {text}
-      </Text>
-    </div>
-  )
-}
-
-Status.propTypes = {
-  type: PropTypes.oneOf(Object.keys(types)).isRequired,
-}
+import { displayCurrency, getStatus } from '../../utils/helpers'
 
 const MyRewards = ({
   myRewards,
@@ -126,14 +92,7 @@ const renderReward = (reward) => {
 
 const renderOneTimeDividend = (reward, amountTokens) => {
   const theme = useTheme()
-  const {
-    description,
-    userRewardAmount,
-    amountToken,
-    dateReference,
-    timeClaimed,
-    endDate
-  } = reward
+  const { description, userRewardAmount, amountToken, dateReference } = reward
   const decimals = amountTokens.find(t => t.symbol === amountToken).decimals
   const displayAmount = (
     <Text color={String(theme.positive)}>
@@ -141,10 +100,8 @@ const renderOneTimeDividend = (reward, amountTokens) => {
     </Text>
   )
   const disbursementDate = dateReference.toDateString()
-  const status = timeClaimed > 0 ? <Status type="claimed" /> : (
-    Date.now() > endDate ? <Status type="ready" /> : <Status type="pending" />
-  )
-  return [ description, disbursementDate, status, displayAmount ]
+  const Status = getStatus(reward)
+  return [ description, disbursementDate, Status, displayAmount ]
 }
 
 const renderRecurringDividend = (reward, amountTokens) => {
@@ -163,22 +120,13 @@ const renderRecurringDividend = (reward, amountTokens) => {
     </Text>
   )
   const disbursementDate = (disbursements[claims] || disbursements[claims-1]).toDateString()
-  const status = claims === disbursements.length ? <Status type="claimed" /> : (
-    Date.now() > disbursements[claims].getTime() ? <Status type="ready" /> :
-      <Status type="pending" />
-  )
-  return [ description, disbursementDate, status, displayAmount ]
+  const Status = getStatus(reward)
+  return [ description, disbursementDate, Status, displayAmount ]
 }
 
 const renderOneTimeMerit = (reward, amountTokens) => {
   const theme = useTheme()
-  const {
-    description,
-    userRewardAmount,
-    amountToken,
-    endDate,
-    timeClaimed
-  } = reward
+  const { description, userRewardAmount, amountToken, endDate } = reward
   const decimals = amountTokens.find(t => t.symbol === amountToken).decimals
   const displayAmount = (
     <Text color={String(theme.positive)}>
@@ -186,10 +134,8 @@ const renderOneTimeMerit = (reward, amountTokens) => {
     </Text>
   )
   const disbursementDate = (new Date(endDate)).toDateString()
-  const status = timeClaimed > 0 ? <Status type="claimed" /> : (
-    Date.now() > endDate ? <Status type="ready" /> : <Status type="pending" />
-  )
-  return [ description, disbursementDate, status, displayAmount ]
+  const Status = getStatus(reward)
+  return [ description, disbursementDate, Status, displayAmount ]
 }
 
 MyRewards.propTypes = {
